requirements.txt
flask
stripe
python-dotenv
python-docx
gunicorn